+++ src/client.rs
@@ -3,6 +3,7 @@ pub mod delete;
 pub mod download;
 pub mod search;
 pub mod upload;
+pub mod users;
 
 pub use auth::{Code, CodeProvider, Token};
 
@@ -10,6 +11,7 @@ use crate::{CenterDevice, ClientCredentials, WithProgress};
 use crate::client::download::Download;
 use crate::client::search::{Search, SearchResult};
 use crate::client::upload::Upload;
+use crate::client::users::{UsersQuery, UsersResult};
 use crate::errors::{Error, ErrorKind, Result};
 
 use failure::Fail;
@@ -82,6 +84,10 @@ impl<'a> CenterDevice for AuthorizedClient<'a> {
 
     fn delete_documents(&self, document_ids: &[&str]) -> Result<()> {
         delete::delete_documents(self, document_ids)
+    }
+
+    fn search_users(&self, users_query: UsersQuery) -> Result<UsersResult> {
+        users::search_users(self, users_query)
     }
 }